Moved formatters to separate folder (#1643)
