+++ core/formatters/word_formatter.talon-list
